+++ pdfalyzer/__init__.py
@@ -124,21 +124,23 @@ def extract_pdf_pages() -> None:
 def extract_pdf_text() -> None:
     """Extract text from a list of file or from all PDF files in a list of directories."""
     args: Namespace = parse_text_extraction_args()
-    console.line()
 
     for file_path in args.files_to_process:
         pdf_file = PdfFile(file_path)
+        console.line()
 
         if args.output_dir:
             txt_file_basename = replace_extension(file_path, 'txt').name
             txt_file_path = args.output_dir.joinpath(txt_file_basename)
             log.warning(f"Extracting '{file_path}'\n        to: '{txt_file_path}'")
 
-            with open(txt_file_path, 'wt') as txt_file:
-                if (extracted_text := pdf_file.extract_text()):
+            if (extracted_text := pdf_file.extract_text(with_page_number_panels=not args.no_page_number_panels)):
+                with open(txt_file_path, 'wt') as txt_file:
                     txt_file.write(extracted_text + "\n")
-                else:
-                    log.warning(f"No text extracted from '{file_path}'...")
+
+                print(extracted_text)
+            else:
+                log.warning(f"No text extracted from '{file_path}'...")
         else:
             pdf_file.print_extracted_text(args.page_range, args.print_as_parsed)
             console.line(2)